# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Restaurant Orders",
    "description" : "Simple API to manage restaurant orders",
    "contact" : {
      "name" : "Sergio Adonis",
      "email" : "sergio.adonis@outlook.com"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/sergioadonis/restaurant-orders-api/v1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
...
    "/restaurants" : {
      "get" : {
        "tags" : [ "Restaurant" ],
        "operationId" : "getRestaurants",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 50
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 0
          }
        }, {
          "name" : "categoryId",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantArrayResult"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Restaurant" ],
        "operationId" : "addRestaurant",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/restaurantBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantObjectResult"
                }
              }
            }
          }
        }
      }
    },
    "/restaurants/{id}" : {
      "get" : {
        "tags" : [ "Restaurant" ],
        "operationId" : "getRestaurantById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantObjectResult"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Restaurant" ],
        "operationId" : "updateRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/restaurantBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantObjectResult"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Restaurant" ],
        "operationId" : "deleteRestaurant",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestaurantObjectResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Restaurant" : {
        "type" : "object",
        "properties" : {
          "categoryId" : {
            "$ref" : "#/components/schemas/ReferenceId"
          },
          "tags" : {
            "$ref" : "#/components/schemas/Tags"
          },
          "contact" : {
            "$ref" : "#/components/schemas/Contact"
          },
          "address" : {
            "$ref" : "#/components/schemas/Address"
          },
          "geoLocation" : {
            "$ref" : "#/components/schemas/GeoLocation"
          },
          "schedules" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Schedule"
            }
          }
        },
        "example" : {
          "address" : {
            "country" : "country",
            "city" : "city",
            "description" : "description",
            "state" : "state",
            "line2" : "line2",
            "line1" : "line1"
          },
          "geoLocation" : {
            "latitud" : 0,
            "longitud" : 6
          },
          "contact" : {
            "whatsapp" : "7788-9900",
            "phoneNumber" : "2222-5555",
            "facebook" : "facebook",
            "email" : "correo@ejemplo.com"
          },
          "schedules" : [ {
            "name" : "name",
            "closeTime" : "closeTime",
            "days" : "days",
            "openTime" : "openTime"
          }, {
            "name" : "name",
            "closeTime" : "closeTime",
            "days" : "days",
            "openTime" : "openTime"
          } ],
          "categoryId" : 123,
          "tags" : [ "Mexican", "Japan", "Favorite" ]
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/CommonProps"
        } ]
      },
...
      "CommonProps" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "readOnly" : true,
            "example" : 123
          },
          "name" : {
            "type" : "string",
            "example" : "Name of object"
          },
          "description" : {
            "type" : "string",
            "example" : "This is a description for this object, could be Markdown sintax."
          },
          "photoUrl" : {
            "type" : "string",
            "example" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9"
          }
        },
        "example" : {
          "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
          "name" : "Name of object",
          "description" : "This is a description for this object, could be Markdown sintax.",
          "id" : 123
        }
      },
      "ReferenceId" : {
        "type" : "integer",
        "format" : "int64",
        "writeOnly" : true,
        "example" : 123
      },
      "Tags" : {
        "type" : "array",
        "example" : [ "Mexican", "Japan", "Favorite" ],
        "items" : {
          "type" : "string"
        }
      },
      "Contact" : {
        "type" : "object",
        "properties" : {
          "phoneNumber" : {
            "type" : "string",
            "example" : "2222-5555"
          },
          "email" : {
            "type" : "string",
            "example" : "correo@ejemplo.com"
          },
          "whatsapp" : {
            "type" : "string",
            "example" : "7788-9900"
          },
          "facebook" : {
            "type" : "string"
          }
        },
        "example" : {
          "whatsapp" : "7788-9900",
          "phoneNumber" : "2222-5555",
          "facebook" : "facebook",
          "email" : "correo@ejemplo.com"
        }
      },
      "Address" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "line1" : {
            "type" : "string"
          },
          "line2" : {
            "type" : "string"
          }
        },
        "example" : {
          "country" : "country",
          "city" : "city",
          "description" : "description",
          "state" : "state",
          "line2" : "line2",
          "line1" : "line1"
        }
      },
      "GeoLocation" : {
        "type" : "object",
        "properties" : {
          "latitud" : {
            "type" : "integer"
          },
          "longitud" : {
            "type" : "integer"
          }
        },
        "example" : {
          "latitud" : 0,
          "longitud" : 6
        }
      },
      "Schedule" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "days" : {
            "type" : "string"
          },
          "openTime" : {
            "type" : "string"
          },
          "closeTime" : {
            "type" : "string"
          }
        },
        "example" : {
          "name" : "name",
          "closeTime" : "closeTime",
          "days" : "days",
          "openTime" : "openTime"
        }
      },
      "Result" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
...
      "RestaurantArrayResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/RestaurantArrayResult_data"
          }
        },
        "example" : {
          "data" : {
            "restaurants" : [ {
              "address" : {
                "country" : "country",
                "city" : "city",
                "description" : "description",
                "state" : "state",
                "line2" : "line2",
                "line1" : "line1"
              },
              "geoLocation" : {
                "latitud" : 0,
                "longitud" : 6
              },
              "contact" : {
                "whatsapp" : "7788-9900",
                "phoneNumber" : "2222-5555",
                "facebook" : "facebook",
                "email" : "correo@ejemplo.com"
              },
              "schedules" : [ {
                "name" : "name",
                "closeTime" : "closeTime",
                "days" : "days",
                "openTime" : "openTime"
              }, {
                "name" : "name",
                "closeTime" : "closeTime",
                "days" : "days",
                "openTime" : "openTime"
              } ],
              "categoryId" : 123,
              "tags" : [ "Mexican", "Japan", "Favorite" ]
            }, {
              "address" : {
                "country" : "country",
                "city" : "city",
                "description" : "description",
                "state" : "state",
                "line2" : "line2",
                "line1" : "line1"
              },
              "geoLocation" : {
                "latitud" : 0,
                "longitud" : 6
              },
              "contact" : {
                "whatsapp" : "7788-9900",
                "phoneNumber" : "2222-5555",
                "facebook" : "facebook",
                "email" : "correo@ejemplo.com"
              },
              "schedules" : [ {
                "name" : "name",
                "closeTime" : "closeTime",
                "days" : "days",
                "openTime" : "openTime"
              }, {
                "name" : "name",
                "closeTime" : "closeTime",
                "days" : "days",
                "openTime" : "openTime"
              } ],
              "categoryId" : 123,
              "tags" : [ "Mexican", "Japan", "Favorite" ]
            } ]
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
      "RestaurantObjectResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/RestaurantObjectResult_data"
          }
        },
        "example" : {
          "data" : {
            "restaurant" : {
              "menus" : [ null, null ],
              "category" : {
                "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
                "name" : "Name of object",
                "description" : "This is a description for this object, could be Markdown sintax.",
                "id" : 123
              },
              "products" : [ null, null ]
            }
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
...
      "RestaurantArrayResult_data" : {
        "type" : "object",
        "properties" : {
          "restaurants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Restaurant"
            }
          }
        },
        "example" : {
          "restaurants" : [ {
            "address" : {
              "country" : "country",
              "city" : "city",
              "description" : "description",
              "state" : "state",
              "line2" : "line2",
              "line1" : "line1"
            },
            "geoLocation" : {
              "latitud" : 0,
              "longitud" : 6
            },
            "contact" : {
              "whatsapp" : "7788-9900",
              "phoneNumber" : "2222-5555",
              "facebook" : "facebook",
              "email" : "correo@ejemplo.com"
            },
            "schedules" : [ {
              "name" : "name",
              "closeTime" : "closeTime",
              "days" : "days",
              "openTime" : "openTime"
            }, {
              "name" : "name",
              "closeTime" : "closeTime",
              "days" : "days",
              "openTime" : "openTime"
            } ],
            "categoryId" : 123,
            "tags" : [ "Mexican", "Japan", "Favorite" ]
          }, {
            "address" : {
              "country" : "country",
              "city" : "city",
              "description" : "description",
              "state" : "state",
              "line2" : "line2",
              "line1" : "line1"
            },
            "geoLocation" : {
              "latitud" : 0,
              "longitud" : 6
            },
            "contact" : {
              "whatsapp" : "7788-9900",
              "phoneNumber" : "2222-5555",
              "facebook" : "facebook",
              "email" : "correo@ejemplo.com"
            },
            "schedules" : [ {
              "name" : "name",
              "closeTime" : "closeTime",
              "days" : "days",
              "openTime" : "openTime"
            }, {
              "name" : "name",
              "closeTime" : "closeTime",
              "days" : "days",
              "openTime" : "openTime"
            } ],
            "categoryId" : 123,
            "tags" : [ "Mexican", "Japan", "Favorite" ]
          } ]
        }
      },
      "RestaurantObjectResult_data_restaurant" : {
        "type" : "object",
        "properties" : {
          "category" : {
            "$ref" : "#/components/schemas/CommonProps"
          },
          "menus" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommonProps"
            }
          },
          "products" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommonProps"
            }
          }
        },
        "example" : {
          "menus" : [ null, null ],
          "category" : {
            "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
            "name" : "Name of object",
            "description" : "This is a description for this object, could be Markdown sintax.",
            "id" : 123
          },
          "products" : [ null, null ]
        }
      },
      "RestaurantObjectResult_data" : {
        "type" : "object",
        "properties" : {
          "restaurant" : {
            "$ref" : "#/components/schemas/RestaurantObjectResult_data_restaurant"
          }
        },
        "example" : {
          "restaurant" : {
            "menus" : [ null, null ],
            "category" : {
              "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
              "name" : "Name of object",
              "description" : "This is a description for this object, could be Markdown sintax.",
              "id" : 123
            },
            "products" : [ null, null ]
          }
        }
      },
...
    },
...
    "requestBodies" : {
...
      "restaurantBody" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Restaurant"
            }
          }
        },
        "required" : true
      },
...
    }
  }
}